Massive refactoring to include event bus. Map now belongs in the movement system Container allows you to fetch a specific system Basic event bus with events (strings) and listeners Game elapsed time is a double, not a float Movement system broadcasts player move for drawer Player is singleton for now

+++ EscapeToEarth/Ecs/Systems/ISystem.cs
@@ -5,6 +5,6 @@ namespace EscapeToEarth.Ecs.Systems
     public interface ISystem
     {
         void Add(Entity e);
-        void Update(float elapsedSeconds);
+        void Update(double elapsedSeconds);
     }
 }
